+++ docs/testing.md
@@ -21,7 +21,7 @@ npx vitest --watch
 
 ## Что покрыто
 
-- Схемы данных и утилиты (`src/lib/*`).
+- Схемы данных и утилиты (`packages/core/src/*`).
 - Геометрия хотспотов и панель добавления (`src/components/*`).
 - Корневой компонент приложения `App`.
