+++ packages/core/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "core",
+  "version": "0.1.0",
+  "type": "module",
+  "main": "src/index.ts",
+  "dependencies": {
+    "zod": "^3.23.8"
+  },
+  "peerDependencies": {
+    "react": "^18.2.0"
+  }
+}
