+++ packages/core/src/dialogSchema.ts
