+++ packages/core/src/imageCache.ts
