+++ packages/core/src/sceneSchema.test.ts
