+++ packages/core/src/sceneSchema.ts
