+++ src/components/DialogEditor.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react"
 import ReactFlow, { Background, Controls, MiniMap, addEdge, Connection, ReactFlowProvider, Node, Edge, useEdgesState, useNodesState } from "reactflow"
 import 'reactflow/dist/style.css'
-import { DialogProject, emptyDialogProject, validateDialogProject } from "@lib/dialogSchema"
-import { loadFileAsText, saveTextFile } from "@lib/utils"
-import { useUndoState } from "@lib/useUndo"
+import { DialogProject, emptyDialogProject, validateDialogProject } from "@core/dialogSchema"
+import { loadFileAsText, saveTextFile } from "@core/utils"
+import { useUndoState } from "@core/useUndo"
 
 interface DialogNodeData {
   label: string
