+++ src/components/ScenesEditor.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useMemo, useRef, useState } from "react"
-import { SceneProject, emptyProject, validateSceneProject, Layer } from "@lib/sceneSchema"
-import { useUndoState } from "@lib/useUndo"
-import type { Point, Hotspot } from "@lib/sceneSchema"
-import { round3, convertProjectCoordsMode } from "@lib/utils"
-import { importSceneProjectFromFile, exportSceneProjectToFile, saveProjectToStorage, loadProjectFromStorage, parseSceneProject } from "@lib/scenePersistence"
+import { SceneProject, emptyProject, validateSceneProject, Layer } from "@core/sceneSchema"
+import { useUndoState } from "@core/useUndo"
+import type { Point, Hotspot } from "@core/sceneSchema"
+import { round3, convertProjectCoordsMode } from "@core/utils"
+import { importSceneProjectFromFile, exportSceneProjectToFile, saveProjectToStorage, loadProjectFromStorage, parseSceneProject } from "@core/scenePersistence"
 import HotspotPanel from "./HotspotPanel"
 import { drawHotspot, hitTestHotspot, translateHotspot, moveVertexTo, setCircleRadius, insertVertex } from "./HotspotShape"
 import LayerPanel from "./LayerPanel"
